dcnv2编译

+++ seg_nas_codes/DCNv2/src/cuda/dcn_v2_im2col_cuda.cu
@@ -6,9 +6,16 @@
 #include <ATen/ATen.h>
 #include <ATen/cuda/CUDAContext.h>
 
-#include <THC/THC.h>
-#include <THC/THCAtomics.cuh>
-#include <THC/THCDeviceUtils.cuh>
+// #include <THC/THC.h>
+// #include <THC/THCAtomics.cuh>
+// #include <THC/THCDeviceUtils.cuh>
+
+// For pytorch 1.11+ 
+// https://github.com/pytorch/pytorch/pull/65472 
+// https://github.com/pytorch/pytorch/pull/66391
+// https://github.com/pytorch/pytorch/pull/65492 
+// MOD by rathaROG 22/03/2022
+
 
 #define CUDA_KERNEL_LOOP(i, n)                          \
   for (int i = blockIdx.x * blockDim.x + threadIdx.x;   \
@@ -25,8 +32,8 @@ inline int GET_BLOCKS(const int N)
 __device__ float dmcn_im2col_bilinear_cuda(const float *bottom_data, const int data_width,
                                       const int height, const int width, float h, float w)
 {
-  int h_low = floor(h);
-  int w_low = floor(w);
+  int h_low = floorf(h);
+  int w_low = floorf(w);
   int h_high = h_low + 1;
   int w_high = w_low + 1;
 
@@ -62,8 +69,8 @@ __device__ float dmcn_get_gradient_weight_cuda(float argmax_h, float argmax_w,
     return 0;
   }
 
-  int argmax_h_low = floor(argmax_h);
-  int argmax_w_low = floor(argmax_w);
+  int argmax_h_low = floorf(argmax_h);
+  int argmax_w_low = floorf(argmax_w);
   int argmax_h_high = argmax_h_low + 1;
   int argmax_w_high = argmax_w_low + 1;
 
@@ -89,8 +96,8 @@ __device__ float dmcn_get_coordinate_weight_cuda(float argmax_h, float argmax_w,
     return 0;
   }
 
-  int argmax_h_low = floor(argmax_h);
-  int argmax_w_low = floor(argmax_w);
+  int argmax_h_low = floorf(argmax_h);
+  int argmax_w_low = floorf(argmax_w);
   int argmax_h_high = argmax_h_low + 1;
   int argmax_w_high = argmax_w_low + 1;
